Net8

+++ src/DH.SignalR/Dtos/NotifyConnectsData.cs
@@ -8,7 +8,7 @@ public class NotifyConnectsData
     /// <summary>
     /// 用户Id
     /// </summary>
-    public Int32 UserId { set; get; }
+    public Int64 UserId { set; get; }
 
     /// <summary>
     /// 连接Id
